fix issue #3
a proposed fix to issue #3 moving init and setters to a seperate file.
