+++ Makefile
@@ -1,5 +1,5 @@
 output:
-	$(CC) -framework Foundation -o notify notify.c
+	$(CC) -framework Foundation -o notify notify.c functions.c
 
 clean:
 	rm -f notify
